+++ Oms-functions-pom/Oms-biz-orderop/src/test/java/com/addressinfo/ChannlServiceEditJunitTest.java
@@ -71,7 +71,7 @@ public class ChannlServiceEditJunitTest extends TestCase {
 	}
 	
 	private OrderGoodsUpdateBean buildUpdateGoods(String orderSn, Long id, String sku, Double TransactionPrice,
-			Double settlementPrice, Double shareBonus, String supplierCode, short number, String depot) {
+			Double settlementPrice, Double shareBonus, String supplierCode, int number, String depot) {
 		if (StringUtil.isTrimEmpty(depot)) {
 			depot = Constant.DETAILS_DEPOT_CODE;
 		}
