+++ Oms-functions-pom/Oms-biz-orderop/src/test/java/com/addressinfo/OrderGoodsEditJunitTest.java
@@ -90,7 +90,7 @@ public class OrderGoodsEditJunitTest extends TestCase {
 	}
 	
 	private OrderGoodsUpdateBean buildUpdateGoods(String orderSn, Long id, String sku, Double TransactionPrice,
-			Double settlementPrice, Double shareBonus, String supplierCode, short number, String depot) {
+			Double settlementPrice, Double shareBonus, String supplierCode, int number, String depot) {
 		if (StringUtil.isTrimEmpty(depot)) {
 			depot = Constant.DETAILS_DEPOT_CODE;
 		}
@@ -111,7 +111,7 @@ public class OrderGoodsEditJunitTest extends TestCase {
 	}
 	
 	private OrderGoodsUpdateBean buildAddGoods(String orderSn, Long id, String sku, Double TransactionPrice,
-			Double settlementPrice, Double shareBonus, String supplierCode, short number, String depot) {
+			Double settlementPrice, Double shareBonus, String supplierCode, int number, String depot) {
 		if (StringUtil.isTrimEmpty(depot)) {
 			depot = Constant.DETAILS_DEPOT_CODE;
 		}
